List the array structure variables

A generic function returning a list with the data structure.

array_vars(
  channel = "635",
  totsamples,
  FG = "",
  BG = "",
  FBG = "",
  blockspersample,
  chip_path = "data/array_data",
  sampleID_path = "data/array_sampleID/",
  mig_prefix = "_first",
  machine = "",
  date_process = ""
)

Arguments

channel

A character indicating the channel that the data was scanned at. It is mostly included in the MFI variable names.

totsamples

A numeric value indicating teh number of samples on a slide.

FG

Optional:A character indicating the name of the foreground variable name. if not specified its created as paste0("F",channel,".Median")

BG

Optional:A character indicating the name of the background variable name. if not specified its created as paste0("B",channel,".Median")

FBG

Optional:A character indicating the name of the foreground - background variable name. if not specified its created as paste0("F",channel,".Median...B",channel)

blockspersample

A numeric value indicating the numer of blocks in a mini-array. The ".gal" file can help in getting this

chip_path

A character indicating the path of the folder location with the array data.

sampleID_path

A character indicating the path of the folder location with the sample identifiers matching the array structure.

mig_prefix

Optional: A character indicating the identifier of an MIG dilution file

machine

Optional:A character indicating the machine used to process the data in the folder

date_process

Optional:A character indicating the date when the samples were processed.

Value

a list of parameters required to process the data

genepix_vars

Examples

## specify the the parameters to process the data
genepix_vars <- array_vars(
## the channel the data was processed in
  channel = "635",
  ## folder where the array data is stored
  chip_path = "data/array_data",
  ## the number of samples per slide or in as single run
  totsamples = 21,
  ## How many blocks each sample occupies
  blockspersample = 2,
  ## folder where the array data samples id files are stored
  sampleID_path = "data/array_sampleID/",
  ## optional
  mig_prefix = "_first",
  machine = 1,
  date_process = "0520"
)
genepix_vars
#> $FG
#> F635.Median
#> 
#> $BG
#> B635.Median
#> 
#> $FBG
#> F635.Median...B635
#> 
#> $paths
#> character(0)
#> 
#> $chip_path
#> [1] "data/array_data"
#> 
#> $sampleID_path
#> [1] "data/array_sampleID/"
#> 
#> $mig_prefix
#> [1] "_first"
#> 
#> $machine
#> [1] 1
#> 
#> $date_process
#> [1] "0520"
#> 
#> $totsamples
#> [1] 21
#> 
#> $blockspersample
#> [1] 2
#> 
#> $mp
#> [1] 1
#> 
#> $dp
#> [1] "0520"
#>