best CV estimation

A function to select the best CV by combining the replicates in duplicates. The function has been build for up to to 3 replicates so far

best_CV_estimation(dataCV, slide_id, lab_replicates, cv_cut_off)

Arguments

dataCV
A data frame

slide_id
A character string containing the identifier of the data frame variable.

lab_replicates
A numeric value indicating the number of lab replicates.

cv_cut_off
a numeric value for the CV cut off. Should be between 0-100

Value

A data frame with the best CV's estimated

Details

Select set of replicates with the best CV

Examples

dataC <- readr::read_csv(system.file("extdata", "dataC.csv", package="protGear"))

#> Rows: 8064 Columns: 11

#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (3): sampleID, antigen, iden
#> dbl (8): sample_array_ID, FMedian, BGMedian, FMedianBG_correct, Block, Colum...

#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.

## this file has 3 lab replicates and the default names
dataCV <- cv_estimation(dataC  ,lab_replicates=3)

#> [1] "The replicates are as expected per sample per antigen"

best_CV_estimation(dataCV,slide_id = "iden", lab_replicates = 3, cv_cut_off = 20)

#> Adding missing grouping variables: `row`

#> # A tibble: 2,457 x 27
#> # Groups:   row [2,457]
#>      row antigen sampleID sample_array_ID   meanX     sdX sdX2_X3 sdX1_X3 sdX1_X2
#>    <int> <chr>   <chr>              <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1     1 AARP    0.0085                11  -10.7     5.13    2.12    4.95    7.07
#>  2     2 AARP    0.0255                 8  -11.3     3.79    4.95    4.24    0.71
#>  3     3 AARP    0.076                  5   -8      11.1     9.9     5.66   15.6 
#>  4     4 AARP    0.23                   2  -12.7     9.61    8.49   13.4     4.95
#>  5     5 AARP    0.73                  19   -6.33   18.9     3.54   21.2    24.8 
#>  6     6 AARP    177.78                 4   58.3   472.    460.    648.    188.  
#>  7     7 AARP    19.75                 10   98     100.    109.    133.     24.0 
#>  8     8 AARP    2.19                  16   -2.33   34.1    43.1     2.83   40.3 
#>  9     9 AARP    533                    1 -520    1312.    264.   1458.   1722.  
#> 10    10 AARP    59.26                  7  395.    394.    457.    505.     48.1 
#> # ... with 2,447 more rows, and 18 more variables: CVX <dbl>, CVX2_X3 <dbl>,
#> #   CVX1_X3 <dbl>, CVX1_X2 <dbl>, cvCat_all <chr>, cvSelected_all <dbl>,
#> #   iden <chr>, 1 <dbl>, 2 <dbl>, 3 <dbl>, x <chr>, meanX1_X2 <dbl>,
#> #   meanX1_X3 <dbl>, meanX2_X3 <dbl>, meanSelected <dbl>, mean_best_CV <dbl>,
#> #   best_CV <dbl>, best_CV_cat <chr>