Read a gpr file to visualize

Read a gpr file to visualize

read_array_visualize(infile)

Arguments

infile
a .gpr file to be used to visualize the expression intensities of the slide spots

Value

a data frame to visualize the background or foreground values

Examples

## Not run:
read_array_visualize(infile = system.file("extdata",
"/array_data/machine1/KK2-06.txt", package="protGear"))

#>       Flags Normalize Autoflag Block Column Row                 Name   ID     X
#>    1:     0         0        0     1      1   1             Landmark  1F1  2310
#>    2:     0         0        0     1      2   1               MSP3.6  1B3  2540
#>    3:     0         0        0     1      3   1           ETRAMP10.2  1H3  2770
#>    4:     0         0        0     1      4   1                 GAMA  1D5  3000
#>    5:     0         0        0     1      5   1                MSRP4  1F5  3230
#>   ---                                                                          
#> 8060:     0         0        0    42      4  24               Buffer 1K12 18630
#> 8061:     0         0        0    42      5  24               Buffer 1M12 18860
#> 8062:     0         0        0    42      6  24               Buffer  1I8 19090
#> 8063:     0         0        0    42      7  24 Commercial Human IgG  1O8 19320
#> 8064:     0         0        0    42      8  24             Landmark 1K10 19550
#>           Y Dia. F635 Median F635 Mean F635 SD F635 CV  B635 B635 Median
#>    1:  1130  150       65535     65330    1317       2  5188        5188
#>    2:  1130  150       13182     15490    8030      51  1652        1652
#>    3:  1130  150        2310      3200    2755      86   899         899
#>    4:  1130  150       18395     18499    7449      40   845         845
#>    5:  1130  150       25786     27006    9322      34   836         836
#>   ---                                                                   
#> 8060: 60450  150         701       719     214      29   574         574
#> 8061: 60450  150        2222      2765    2250      81   685         685
#> 8062: 60450  150       12219     14067   12815      91  1412        1412
#> 8063: 60450  150       65535     65535       0       0 26861       26861
#> 8064: 60450  150       65535     60763   10793      17 27943       27943
#>       B635 Mean B635 SD B635 CV % > B635+1SD % > B635+2SD F635 % Sat.
#>    1:     13457   19832     147          100          100          96
#>    2:     10216   19396     189           18            2           0
#>    3:      1446    2690     186           24            6           0
#>    4:      1014     903      89          100           99           0
#>    5:      1292    2408     186           99           97           0
#>   ---                                                                
#> 8060:       712     941     132            1            0           0
#> 8061:       791     443      56           71           65           0
#> 8062:      4712    9336     198           55           29           0
#> 8063:     29706   23322      78          100            0         100
#> 8064:     32939   18291      55           87           79          79
#>       F635 Total Intensity SNR 635 Ratio of Medians (635/594)
#>    1:             10191549   2.616                      1e+05
#>    2:              2416423   0.272                      1e+05
#>    3:               499274   0.652                      1e+05
#>    4:              2885913  19.363                      1e+05
#>    5:              4212925  10.679                      1e+05
#>   ---                                                        
#> 8060:               112241   0.007                      1e+05
#> 8061:               431327   4.456                      1e+05
#> 8062:              2194514   1.002                      1e+05
#> 8063:             10223460   1.536                      1e+05
#> 8064:              9479001   1.521                      1e+05
#>       Ratio of Means (635/594) Median of Ratios (635/594)
#>    1:                    1e+05                          0
#>    2:                    1e+05                          0
#>    3:                    1e+05                          0
#>    4:                    1e+05                          0
#>    5:                    1e+05                          0
#>   ---                                                    
#> 8060:                    1e+05                          0
#> 8061:                    1e+05                          0
#> 8062:                    1e+05                          0
#> 8063:                    1e+05                          0
#> 8064:                    1e+05                          0
#>       Mean of Ratios (635/594) Ratios SD (635/594) Rgn Ratio (635/594)
#>    1:                        0                   0                   0
#>    2:                        0                   0                   0
#>    3:                        0                   0                   0
#>    4:                        0                   0                   0
#>    5:                        0                   0                   0
#>   ---                                                                 
#> 8060:                        0                   0                   0
#> 8061:                        0                   0                   0
#> 8062:                        0                   0                   0
#> 8063:                        0                   0                   0
#> 8064:                        0                   0                   0
#>       Rgn R2 (635/594) Log Ratio (635/594) F635 Median - B635 F635 Mean - B635
#>    1:                0              16.610              60347            60142
#>    2:                0              16.610              11530            13838
#>    3:                0              16.610               1411             2301
#>    4:                0              16.610              17550            17654
#>    5:                0              16.610              24950            26170
#>   ---                                                                         
#> 8060:                0              16.610                127              145
#> 8061:                0              16.610               1537             2080
#> 8062:                0              16.610              10807            12655
#> 8063:                0              16.610              38674            38674
#> 8064:                0              16.610              37592            32820
#>       Sum of Medians (635/594) Sum of Means (635/594) F Pixels B Pixels
#>    1:                    60347                  60142      156     1072
#>    2:                    11530                  13838      156      972
#>    3:                     1411                   2301      156      972
#>    4:                    17550                  17654      156      972
#>    5:                    24950                  26170      156      972
#>   ---                                                                  
#> 8060:                      127                    145      156      972
#> 8061:                     1537                   2080      156      972
#> 8062:                    10807                  12655      156      972
#> 8063:                    38674                  38674      156      972
#> 8064:                    37592                  32820      156     1072
#>       Circularity Index
#>    1:         100     1
#>    2:         100     2
#>    3:         100     3
#>    4:         100     4
#>    5:         100     5
#>   ---                  
#> 8060:         100  8060
#> 8061:         100  8061
#> 8062:         100  8062
#> 8063:         100  8063
#> 8064:         100  8064

## End(Not run)