bg_correct

A generic function to perform background correction.

bg_correct(iden, Data1, genepix_vars, method = "subtract_local")

Arguments

iden
A character indicating the name of the object to be used under Data1

Data1
A data frame with sample identifiers merged with micro array data.

genepix_vars
A list of specific definitions of the experiment design. See array_vars.

method
a description of the background correction to be used. Possible values are "none","subtract_local", "subtract_global","movingmin_bg","minimum_half","edwards" or "normexp". The default is "subtract_local".

Value

A data frame with background corrected data

Details

Background correction

The function implements background correction methods developed by backgroundCorrect. But the minimum_half and movingmin_bg uses the block of the protein array as the grid. If method="movingmin_bg" the minimum background value within a block is subtracted. If method="minimum_half" then any intensity which is negative after background subtraction is reset to be equal to half the minimum positive value in a block. If method="movingmin_value" then any intensity which is negative after background subtraction is reset to the minimum positive value in a block. For edwards we implement a similar algorithm with limma::backgroundCorrect(method="edwards") and for 'normexp' we use the saddle-point approximation to maximum likelihood, backgroundCorrect for more details.

Examples

## Not run:
genepix_vars <- array_vars(
  channel = "635",
  chip_path = system.file("extdata", "array_data/machine1/", package="protGear"),
  totsamples = 21,
  blockspersample = 2,
  mig_prefix = "_first",
  machine = 1,
  ## optional
  date_process = "0520"
)
raw_df <- readr::read_csv(system.file("extdata", "Data1_bg_sample.csv", package="protGear"))

#> Rows: 8064 Columns: 15

#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr  (3): Name, ID, sampleID
#> dbl (12): Block, Column, Row, F635 Median, B635, B635 Median, SNR 635, F635 ...

#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.

bg_correct(iden="iden",
Data1 = raw_df,
genepix_vars = genepix_vars, method="subtract_local"
)

#> # A tibble: 8,064 x 10
#>    sampleID sample_array_ID antigen    FMedian BGMedian FMedianBG_correct Block
#>    <chr>              <dbl> <chr>        <dbl>    <dbl>             <dbl> <dbl>
#>  1 S_g1                   1 Landmark     65535     5188             60347     1
#>  2 S_g1                   1 MSP3.6       13182     1652             11530     1
#>  3 S_g1                   1 ETRAMP10.2    2310      899              1411     1
#>  4 S_g1                   1 GAMA         18395      845             17550     1
#>  5 S_g1                   1 MSRP4        25786      836             24950     1
#>  6 S_g1                   1 MSP4         15503      790             14713     1
#>  7 S_g1                   1 RAMA          7602      753              6849     1
#>  8 S_g1                   1 ASP          17234      754             16480     1
#>  9 S_g1                   1 Landmark     65535     5024             60511     1
#> 10 S_g1                   1 MSP3.6       12911     2590             10321     1
#> # ... with 8,054 more rows, and 3 more variables: Column <dbl>, Row <dbl>,
#> #   replicate <int>

## End(Not run)