bg_correct

A generic function to perform background correction.

bg_correct(iden, Data1, genepix_vars, method = "subtract_local")

Arguments

iden

A character indicating the name of the object to be used under Data1

Data1

A data frame with sample identifiers merged with micro array data.

genepix_vars

A list of specific definitions of the experiment design. See array_vars.

method

a description of the background correction to be used. Possible values are "none","subtract_local", "subtract_global","movingmin_bg","minimum_half","edwards" or "normexp". The default is "subtract_local".

Value

A data frame with background corrected data

Details

Background correction

The function implements background correction methods developed by backgroundCorrect. But the minimum_half and movingmin_bg uses the block of the protein array as the grid. If method="movingmin_bg" the minimum background value within a block is subtracted. If method="minimum_half" then any intensity which is negative after background subtraction is reset to be equal to half the minimum positive value in a block. If method="movingmin_value" then any intensity which is negative after background subtraction is reset to the minimum positive value in a block. For edwards we implement a similar algorithm with limma::backgroundCorrect(method="edwards") and for 'normexp' we use the saddle-point approximation to maximum likelihood, backgroundCorrect for more details.