Read a gpr file to visualize

Read a gpr file to visualize

read_array_visualize(infile)

Arguments

infile

a .gpr file to be used to visualize the expression intensities of the slide spots

Value

a data frame to visualize the background or foreground values

Examples

## Not run:
read_array_visualize(infile = system.file("extdata",
"/array_data/machine1/KK2-06.txt", package="protGear"))
#>       Flags Normalize Autoflag Block Column Row                 Name   ID     X
#>    1:     0         0        0     1      1   1             Landmark  1F1  2130
#>    2:     0         0        0     1      2   1               MSP3.6  1B3  2360
#>    3:     0         0        0     1      3   1           ETRAMP10.2  1H3  2590
#>    4:     0         0        0     1      4   1                 GAMA  1D5  2820
#>    5:     0         0        0     1      5   1                MSRP4  1F5  3050
#>   ---                                                                          
#> 8060:     0         0        0    42      4  24               Buffer 1K12 18490
#> 8061:     0         0        0    42      5  24               Buffer 1M12 18720
#> 8062:     0         0        0    42      6  24               Buffer  1I8 18950
#> 8063:     0         0        0    42      7  24 Commercial Human IgG  1O8 19180
#> 8064:     0         0        0    42      8  24             Landmark 1K10 19410
#>           Y F635 Median  B635 B635 Median SNR 635 F635 Median - B635 Index
#>    1:  1150       65535  2443        2443   4.575              63092     1
#>    2:  1150        5488  1126        1126   0.153               4362     2
#>    3:  1150        6350  1026        1026  11.718               5324     3
#>    4:  1150       15768  1059        1059  10.836              14709     4
#>    5:  1150       65535  1413        1413   9.025              64122     5
#>   ---                                                                     
#> 8060: 60490         820   690         690   0.107                130  8060
#> 8061: 60490        1168   737         737   0.423                431  8061
#> 8062: 60490        2383  1125        1125   0.223               1258  8062
#> 8063: 60490       65535 19685       19685   1.936              45850  8063
#> 8064: 60490       65535 18134       18134   1.564              47401  8064
## End(Not run)